--- Assets/Game_End_UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d9ec4756a9a94d24da3b8a6d9eed0328
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
