--- Assets/Game_End_UI/Audio.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 06ec1b05720d9cb44aa6df00eff98531
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
