--- Assets/Game_End_UI/Scenes/LostScreen.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1dd94b4ae56904d4cad129d943415140
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
