--- Assets/Game_End_UI/Scenes/WInScereen 1.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1660a6e3606255d41832b269ec83ecea
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
