+++ Assets/MainPackage/Game_End_UI/Audio/female-horror-ghost-sound-2-vol-002-140412 (mp3cut.net).mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 2b9a4d056d89c224b81c9a73f7156abe
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
