+++ Assets/MainPackage/Game_End_UI/Scenes/LostScreen.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 038584a58f21e314baa1dc35b9e93e13
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
