+++ Assets/MainPackage/Scripts/NPC.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.AI;
+
+public class NPC : MonoBehaviour
+{
+    // Start is called before the first frame update
+    public Transform Follow;
+    NavMeshAgent agent;
+    void Start()
+    {
+        agent = GetComponent<NavMeshAgent>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        agent.destination = Follow.position;
+    }
+}
