+++ Assets/MainPackage/Scripts/NPC.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a16f7513caa46374e8647a2cc3485299
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
