+++ Assets/MainPackage/Scripts/NPC_Collision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ceaef1bd3f5071f4fbeb531aabcaa493
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
