--- Assets/NPC.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9a2da7ae8a490484fbc3993652732a17
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
